notecard-forwarder/requirements.txt
python-periphery==2.0.1
filelock==3.0.12
note-python==1.2.1